requirements.txt
streamlit
spacy
nltk
scikit-learn
sentence-transformers
pdfminer.six
python-docx
PyMuPDF
openai
